requirements.txt
streamlit==1.32.2
google-generativeai==0.8.3
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==0.8.11
beautifulsoup4==4.12.3
striprtf==0.0.26
